+++ backend/src/api/report/content-types/report/schema.json
@@ -34,6 +34,9 @@
         "bug"
       ],
       "required": true
+    },
+    "time": {
+      "type": "datetime"
     }
   }
 }
